Add files via upload
The objective of this project is to model the constituents of the energy balance at the surface of the earth, for different surface types and better understand the how the constituents change depending on surface type. A linear regression analysis is used to developed models of sensible and latent heat as a function of net radiation. This will allow for the estimation of these terms by only measuring net radiation. Meteorological instrumentation was deployed at the UCR’s Citrus Grove (Citrus) and CE-CERT site. Instruments at the grove were deployed above a large area covered with dirt and orange trees while instruments at the CE-CERT site were above a concrete surface with few short buildings. Previous work and the results of this project show that net radiation is a good predictor of sensible, latent, and ground flux terms. A linear regression analysis of the short and long wave terms of the overall net radiation and the gradients of these
terms, my model is more accurate than Grimmond’s Ground Model and the Penman-Monteith method.